fw_info: Extract fw_info struct definition to its own header
Extract the fw_info struct definition to its own header.
This makes it possible to use the struct header in a non-zephyr
build system without having to include all the other zephyr definitions.

NCSDK-14890
